requirements.txt
pytest==8.2.0
pytest-playwright==0.5.1
playwright==1.45.0
python-dotenv==1.0.1